# packages/api-client/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Elton Decor Foundation API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/products": {
      "get": {
        "tags": [
          "catalog"
        ],
        "summary": "Products",
        "operationId": "products_api_v1_products_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Q"
            }
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "title_asc",
                "price_asc",
                "price_desc"
              ],
              "type": "string",
              "default": "title_asc",
              "title": "Sort"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductPage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/products/by-slug/{slug}": {
      "get": {
        "tags": [
          "catalog"
        ],
        "summary": "By Slug",
        "operationId": "by_slug_api_v1_products_by_slug__slug__get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/products/{id}": {
      "get": {
        "tags": [
          "catalog"
        ],
        "summary": "By Id",
        "operationId": "by_id_api_v1_products__id__get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/collections/{id}": {
      "get": {
        "tags": [
          "catalog"
        ],
        "summary": "Collection",
        "operationId": "collection_api_v1_collections__id__get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/guest-sessions": {
      "post": {
        "tags": [
          "guest"
        ],
        "summary": "Guest Sessions",
        "operationId": "guest_sessions_api_v1_guest_sessions_post",
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionDTO"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionDTO"
                }
              }
            }
          },
          "403": {
            "description": "Origin rejected"
          },
          "429": {
            "description": "Rate limited"
          }
        }
      }
    },
...
    "/api/v1/draft-quotes": {
      "post": {
        "tags": [
          "local drafts"
        ],
        "summary": "Create Quote",
        "operationId": "create_quote_api_v1_draft_quotes_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuoteCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DraftQuoteDTO"
                }
              }
            }
          },
          "409": {
            "description": "Cart changed"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/checkout-drafts": {
      "post": {
        "tags": [
          "local drafts"
        ],
        "summary": "Save Draft",
        "operationId": "save_draft_api_v1_checkout_drafts_post",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DraftSaveCommand"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DraftDTO"
                }
              }
            }
          },
          "409": {
            "description": "Idempotency, cart or quote conflict"
          },
          "410": {
            "description": "Quote expired"
          },
          "428": {
            "description": "Idempotency-Key required"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/checkout-drafts/{draft_id}": {
      "get": {
        "tags": [
          "local drafts"
        ],
        "summary": "Get Draft",
        "operationId": "get_draft_api_v1_checkout_drafts__draft_id__get",
        "parameters": [
          {
            "name": "draft_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Draft Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DraftDTO"
                }
              }
            }
          },
          "404": {
            "description": "Unknown or foreign draft"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/session": {
      "get": {
        "tags": [
          "admin foundation"
        ],
        "summary": "Admin Session",
        "operationId": "admin_session_api_v1_admin_session_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminSessionDTO"
                }
              }
            }
          },
          "401": {
            "description": "Admin session required"
          },
          "403": {
            "description": "Origin rejected"
          }
        }
      }
    },
...
    "/api/v1/admin/checkout-drafts": {
      "get": {
        "tags": [
          "admin foundation"
        ],
        "summary": "List Drafts",
        "operationId": "list_drafts_api_v1_admin_checkout_drafts_get",
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 512
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "name": "created_from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Created From"
            }
          },
          {
            "name": "created_to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Created To"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DraftPageDTO"
                }
              }
            }
          },
          "401": {
            "description": "Admin session required"
          },
          "403": {
            "description": "Permission denied"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AdminSessionDTO": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "permissions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Permissions"
          },
          "csrf_token": {
            "type": "string",
            "title": "Csrf Token"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "email",
          "permissions",
          "csrf_token",
          "expires_at"
        ],
        "title": "AdminSessionDTO"
      },
...
      "CollectionDetail": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ProductSummary"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "slug",
          "title",
          "items"
        ],
        "title": "CollectionDetail"
      },
...
      "DraftDTO": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/DraftLineSnapshot"
            },
            "type": "array",
            "title": "Items"
          },
          "goods_total": {
            "$ref": "#/components/schemas/Money"
          },
          "delivery": {
            "$ref": "#/components/schemas/CartDelivery"
          },
          "payable_total": {
            "type": "null",
            "title": "Payable Total"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "state": {
            "type": "string",
            "const": "saved",
            "title": "State",
            "default": "saved"
          },
          "channel": {
            "type": "string",
            "const": "site",
            "title": "Channel",
            "default": "site"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "contact": {
            "$ref": "#/components/schemas/ContactDemo"
          },
          "address": {
            "$ref": "#/components/schemas/AddressDemo"
          },
          "capabilities": {
            "$ref": "#/components/schemas/DraftCapabilities"
          },
          "message": {
            "type": "string",
            "const": "Заявка сохранена; оплата и доставка пока не подключены",
            "title": "Message",
            "default": "Заявка сохранена; оплата и доставка пока не подключены"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items",
          "goods_total",
          "id",
          "created_at",
          "contact",
          "address"
        ],
        "title": "DraftDTO"
      },
...
      "DraftPageDTO": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AdminDraftDTO"
            },
            "type": "array",
            "title": "Items"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items"
        ],
        "title": "DraftPageDTO"
      },
      "DraftQuoteDTO": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/DraftLineSnapshot"
            },
            "type": "array",
            "title": "Items"
          },
          "goods_total": {
            "$ref": "#/components/schemas/Money"
          },
          "delivery": {
            "$ref": "#/components/schemas/CartDelivery"
          },
          "payable_total": {
            "type": "null",
            "title": "Payable Total"
          },
          "quote_id": {
            "type": "string",
            "format": "uuid",
            "title": "Quote Id"
          },
          "cart_version": {
            "type": "integer",
            "title": "Cart Version"
          },
          "state": {
            "type": "string",
            "const": "valid",
            "title": "State",
            "default": "valid"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          },
          "capabilities": {
            "$ref": "#/components/schemas/QuoteCapabilities"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items",
          "goods_total",
          "quote_id",
          "cart_version",
          "expires_at"
        ],
        "title": "DraftQuoteDTO"
      },
      "DraftSaveCommand": {
        "properties": {
          "quote_id": {
            "type": "string",
            "format": "uuid",
            "title": "Quote Id"
          },
          "contact": {
            "$ref": "#/components/schemas/ContactDemo"
          },
          "address": {
            "$ref": "#/components/schemas/AddressDemo"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "quote_id",
          "contact",
          "address"
        ],
        "title": "DraftSaveCommand"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Money": {
        "properties": {
          "amount_minor": {
            "type": "string",
            "pattern": "^(0|[1-9][0-9]*)$",
            "title": "Amount Minor"
          },
          "currency": {
            "type": "string",
            "const": "RUB",
            "title": "Currency",
            "default": "RUB"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "amount_minor"
        ],
        "title": "Money"
      },
...
      "ProductDetail": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "sku": {
            "type": "string",
            "title": "Sku"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "type": {
            "type": "string",
            "enum": [
              "single",
              "bundle"
            ],
            "title": "Type"
          },
          "price": {
            "$ref": "#/components/schemas/Money"
          },
          "original_price": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Money"
              },
              {
                "type": "null"
              }
            ]
          },
          "thumbnail": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MediaDTO"
              },
              {
                "type": "null"
              }
            ]
          },
          "availability": {
            "$ref": "#/components/schemas/Availability"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "categories": {
            "items": {
              "$ref": "#/components/schemas/CategoryDTO"
            },
            "type": "array",
            "title": "Categories"
          },
          "attributes": {
            "additionalProperties": true,
            "type": "object",
            "title": "Attributes"
          },
          "seo": {
            "additionalProperties": true,
            "type": "object",
            "title": "Seo"
          },
          "media": {
            "items": {
              "$ref": "#/components/schemas/MediaDTO"
            },
            "type": "array",
            "title": "Media"
          },
          "related_product_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Related Product Ids"
          },
          "bundle": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BundleDTO"
              },
              {
                "type": "null"
              }
            ]
          },
          "delivery": {
            "$ref": "#/components/schemas/CatalogDelivery"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "sku",
          "slug",
          "title",
          "type",
          "price",
          "version",
          "description",
          "categories",
          "attributes",
          "seo",
          "media",
          "related_product_ids"
        ],
        "title": "ProductDetail"
      },
      "ProductPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ProductSummary"
            },
            "type": "array",
            "title": "Items"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items"
        ],
        "title": "ProductPage"
      },
...
      "QuoteCommand": {
        "properties": {
          "cart_version": {
            "type": "integer",
            "maximum": 9.223372036854776e+18,
            "minimum": 1.0,
            "title": "Cart Version"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "cart_version"
        ],
        "title": "QuoteCommand"
      },
      "SessionDTO": {
        "properties": {
          "csrf_token": {
            "type": "string",
            "title": "Csrf Token"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "csrf_token",
          "expires_at"
        ],
        "title": "SessionDTO"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}